requirements.txt
PyYAML>=3.1,<3.90
jsonschema>=3.0,<3.9
yattag>=1.11.1,<2.0
matplotlib>=2.0.2,<3.9
numpy>=1.12.1
hypothesis>=4.18.0,<5.0.0
